Moved entry-point pages to entry/ folder.
