+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Main from './components/Main'
-import './components/Main.css'
+import Main from './components/entry/Main'
+import './components/entry/Main.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { FirebaseContext, Firebase } from './firebase'
